Move effects and related states to dedicated component hook
